+++ CHANGELOG.md
@@ -84,6 +84,7 @@
 - Enh #85: Add `HasFontFamily` trait and corresponding tests for managing SVG `font-family` attribute (@terabytesoftw)
 - Bug #86: Standardize all documentation in related classes (@terabytesoftw)
 - Enh #87: Add `HasFontSize` trait and corresponding tests for managing SVG `font-size` attribute (@terabytesoftw)
+- Enh #88: Add `HasFontStyle` trait and corresponding tests for managing SVG `font-style` attribute (@terabytesoftw)
 
 ## 0.2.0 March 31, 2024